Move example app to separate file
